Create README.md
The project is about the analysis of Indian premier league (IPL) cricket match data through  shell scripting  and  Linux command-line tools. It goes through the transfer of raw  CSV data, to cleanup and extract useful statistics and provide summaries like the best performing teams, most common tournament locations and the effect of winning the toss. 
Using only the most common tools of the standard-issue Linux operating system (`awk`, `grep`, `cut`, `sort`, and `uniq`), the project illustrates how text-data processing might be powerful even in the absence of bulky data analytics tools. The method facilitates efficiency, portability, and reusability resulting in an approach that is suitable to educational and lightweight analysis of data.